커뮤니티 컨트리 lsm 컨트리로 연동
