--- src/main/java/com/kidmily/algoga_server/country/application/usecase/CountryQueryUseCase.java
@@ -1,10 +1,0 @@
-package com.kidmily.algoga_server.country.application.usecase;
-
-import com.kidmily.algoga_server.country.presentation.api.response.CountryResponse;
-
-import java.util.List;
-
-public interface CountryQueryUseCase {
-
-    List<CountryResponse> getActiveCountries();
-}
