--- src/main/java/com/kidmily/algoga_server/country/presentation/advice/CountryExceptionAdvice.java
@@ -1,29 +1,0 @@
-package com.kidmily.algoga_server.country.presentation.advice;
-
-import com.kidmily.algoga_server.global.common.api.response.ErrorResponse;
-import com.kidmily.algoga_server.global.exception.BaseErrorCode;
-import com.kidmily.algoga_server.global.exception.BusinessException;
-import com.kidmily.algoga_server.global.exception.CommonExceptionAdvice;
-import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice(basePackages = "com.kidmily.algoga_server.country.presentation.api")
-public class CountryExceptionAdvice implements CommonExceptionAdvice {
-
-    @Override
-    public Logger getLogger() {
-        return log;
-    }
-
-    @ExceptionHandler(BusinessException.class)
-    @Override
-    public ResponseEntity<ErrorResponse> handleBusinessException(BusinessException e) {
-        BaseErrorCode errorCode = e.getErrorCode();
-        log.warn("[Business Exception] code: {}, message: {}", errorCode.getCode(), e.getMessage());
-        return CommonExceptionAdvice.super.handleBusinessException(e);
-    }
-}
