+++ src/main/java/com/kidmily/algoga_server/lms/domain/repository/CountryRepository.java
@@ -1,5 +1,7 @@
-package com.kidmily.algoga_server.country.domain.repository;
-import com.kidmily.algoga_server.country.domain.model.Country;
+package com.kidmily.algoga_server.lms.domain.repository;
+
+import com.kidmily.algoga_server.lms.domain.model.Country;
+
 import java.util.List;
 import java.util.Optional;
