+++ src/main/java/com/kidmily/algoga_server/lms/infrastructure/persistence/CountryRepositoryAdapter.java
@@ -1,8 +1,9 @@
-package com.kidmily.algoga_server.country.infrastructure.persistence;
+package com.kidmily.algoga_server.lms.infrastructure.persistence;
 
-import com.kidmily.algoga_server.country.domain.model.Country;
-import com.kidmily.algoga_server.country.domain.repository.CountryRepository;
-import com.kidmily.algoga_server.country.infrastructure.mapper.CountryMapper;
+import com.kidmily.algoga_server.lms.domain.model.Country;
+import com.kidmily.algoga_server.lms.domain.repository.CountryRepository;
+import com.kidmily.algoga_server.lms.infrastructure.mapper.CountryMapper;
+import com.kidmily.algoga_server.lms.infrastructure.persistence.repository.SpringDataCountryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -18,11 +19,12 @@ public class CountryRepositoryAdapter implements CountryRepository {
 
     @Override
     public List<Country> findAllActive() {
-        return springDataRepository.findByIsActiveTrue()
+        return springDataRepository.findByActiveTrue()  // IsActive → Active로 변경
                 .stream()
                 .map(countryMapper::toDomain)
                 .toList();
     }
+
     @Override
     public Optional<Country> findById(Long id) {
         return springDataRepository.findById(id)
