+++ src/main/java/com/kidmily/algoga_server/lms/infrastructure/persistence/repository/SpringDataCountryRepository.java
@@ -1,0 +1,10 @@
+package com.kidmily.algoga_server.lms.infrastructure.persistence.repository;
+
+import com.kidmily.algoga_server.lms.infrastructure.persistence.entity.CountryJpaEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface SpringDataCountryRepository extends JpaRepository<CountryJpaEntity, Long> {
+    List<CountryJpaEntity> findByActiveTrue();
+}
